Fix code review issues: correct SftpAccount model and migration comments

+++ database/migrations/2024_12_01_000008_create_web_features_tables.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('custom_error_pages', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('domain_id');
-            $table->integer('error_code'); // 400, 401, 403, 404, 500, etc.
+            $table->unsignedSmallInteger('error_code'); // 100-599 HTTP status codes
             $table->text('custom_content')->nullable(); // HTML content
             $table->string('custom_file_path')->nullable(); // Path to custom error file
             $table->boolean('is_active')->default(true);
